Move controller to it's own package, it's not part of the registry.
